Fix ci

+++ packages/vcpkg/neko-function/portfile.cmake
@@ -25,7 +25,11 @@ vcpkg_cmake_configure(
 
 vcpkg_cmake_install()
 vcpkg_cmake_config_fixup(CONFIG_PATH lib/cmake/NekoFunction PACKAGE_NAME nekofunction)
-file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/lib")
+
+# Only remove lib directory if both archive and hash features are disabled (header-only mode)
+if(NOT "archive" IN_LIST FEATURES AND NOT "hash" IN_LIST FEATURES)
+    file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/lib")
+endif()
 
 vcpkg_install_copyright(FILE_LIST "${SOURCE_PATH}/LICENSE")
